# Optimization for gridwise group norm