… test/layernorm/test_groupnorm_fp16.cpp