Optimize group norm kernel:
1. coalesce load for x/y/gamma/beta.
2. move a sqrt division computation into outer loop.
3. test a instance with 1024 threads

Gridwise layernorm is modified. Test is in progress. Optimization part can be reviewed. 

On MI100:
./bin/example_groupnorm_sigmoid_fp16 2 32 32 32 30
Before this PR:
```
launch_and_time_kernel: grid_dim {64, 1, 1}, block_dim {1024, 1, 1}
Warm up 1 time
Start running 10 times...
Perf: 0.0767513 ms, 102.515 GB/s, DeviceLayernormImpl<1024,M_C1_S1,K_C1024_S2,XYSrcVectorDim_1,VectorSize_X2_Gamma2_Beta2_Y2>
```
This PR:
```
launch_and_time_kernel: grid_dim {64, 1, 1}, block_dim {1024, 1, 1}
Warm up 1 time
Start running 10 times...
Perf: 0.0585594 ms, 134.362 GB/s, DeviceLayernormImpl<1024,M_C1_S1,K_C1024_S32,XYSrcVectorDim_1,VectorSize_X2_Gamma2_Beta2_Y2>
```